+++ packages/cervel/src/find-entrypoint.ts
@@ -2,9 +2,16 @@ import { readFile } from 'fs/promises';
 import { join } from 'path';
 import { existsSync } from 'fs';
 
-const frameworks = ['express', 'hono'];
+const frameworks = [
+  'express',
+  'hono',
+  'elysia',
+  'fastify',
+  '@nestjs/core',
+  'h3',
+];
 
-const entrypointFilenames = ['app', 'index', 'server'];
+const entrypointFilenames = ['app', 'index', 'server', 'main'];
 
 const entrypointExtensions = ['js', 'cjs', 'mjs', 'ts', 'cts', 'mts'];