+++ packages/node/src/build.ts
@@ -24,6 +24,7 @@ import {
   runNpmInstall,
   runPackageJsonScript,
   getNodeVersion,
+  getSpawnOptions,
   debug,
   isSymbolicLink,
   walkParentDirs,
@@ -81,6 +82,7 @@ async function downloadInstallAndBundle({
     config,
     meta
   );
+  const spawnOpts = getSpawnOptions(meta, nodeVersion);
 
   const {
     cliType,
@@ -89,11 +91,11 @@ async function downloadInstallAndBundle({
     turboSupportsCorepackHome,
   } = await scanParentDirs(entrypointFsDirname, true);
 
-  const spawnEnv = getEnvForPackageManager({
+  spawnOpts.env = getEnvForPackageManager({
     cliType,
     lockfileVersion,
     packageJsonPackageManager,
-    env: process.env,
+    env: spawnOpts.env || {},
     turboSupportsCorepackHome,
     projectCreatedAt: config.projectSettings?.createdAt,
   });
@@ -103,7 +105,7 @@ async function downloadInstallAndBundle({
     if (installCommand.trim()) {
       console.log(`Running "install" command: \`${installCommand}\`...`);
       await execCommand(installCommand, {
-        env: spawnEnv,
+        ...spawnOpts,
         cwd: entrypointFsDirname,
       });
     } else {
@@ -113,13 +115,13 @@ async function downloadInstallAndBundle({
     await runNpmInstall(
       entrypointFsDirname,
       [],
-      { env: spawnEnv },
+      spawnOpts,
       meta,
       config.projectSettings?.createdAt
     );
   }
   const entrypointPath = downloadedFiles[entrypoint].fsPath;
-  return { entrypointPath, entrypointFsDirname, nodeVersion, spawnEnv };
+  return { entrypointPath, entrypointFsDirname, nodeVersion, spawnOpts };
 }
 
 function renameTStoJS(path: string) {
@@ -145,7 +147,8 @@ async function compile(
   config: Config,
   meta: Meta,
   nodeVersion: NodeVersion,
-  isEdgeFunction: boolean
+  isEdgeFunction: boolean,
+  useTypescript5 = false
 ): Promise<{
   preparedFiles: Files;
   shouldAddSourcemapSupport: boolean;
@@ -187,6 +190,7 @@ async function compile(
         project: path, // Resolve tsconfig.json from entrypoint dir
         files: true, // Include all files such as global `.d.ts`
         nodeVersionMajor: nodeVersion.major,
+        useTypescript5,
       });
     }
     const { code, map } = tsCompile(source, path);
@@ -417,6 +421,7 @@ export const build = async ({
   meta = {},
   considerBuildCommand = false,
   entrypointCallback,
+  checks = () => {},
 }: Parameters<BuildV3>[0] & {
   shim?: (handler: string) => string;
   useWebApi?: boolean;
@@ -426,6 +431,7 @@ export const build = async ({
    * from files that may have been created by the build script.
    */
   entrypointCallback?: () => Promise<string>;
+  checks?: (project: { config: Config; isBun: boolean }) => void;
 }): Promise<BuildResultV3> => {
   const baseDir = repoRootPath || workPath;
   const awsLambdaHandler = getAWSLambdaHandler(entrypoint, config);
@@ -434,7 +440,7 @@ export const build = async ({
     entrypointPath: _entrypointPath,
     entrypointFsDirname,
     nodeVersion,
-    spawnEnv,
+    spawnOpts,
   } = await downloadInstallAndBundle({
     files,
     entrypoint,
@@ -453,11 +459,13 @@ export const build = async ({
   // primary builder
   if (projectBuildCommand && considerBuildCommand) {
     await execCommand(projectBuildCommand, {
+      ...spawnOpts,
+
       // Yarn v2 PnP mode may be activated, so force
       // "node-modules" linker style
       env: {
         YARN_NODE_LINKER: 'node-modules',
-        ...spawnEnv,
+        ...spawnOpts.env,
       },
 
       cwd: workPath,
@@ -470,7 +478,7 @@ export const build = async ({
     await runPackageJsonScript(
       entrypointFsDirname,
       possibleScripts,
-      { env: spawnEnv },
+      spawnOpts,
       config.projectSettings?.createdAt
     );
   }
@@ -506,6 +514,13 @@ export const build = async ({
     isEdgeFunction = isEdgeRuntime(runtime);
   }
 
+  checks({
+    config,
+    isBun: isBunVersion(nodeVersion),
+  });
+
+  // Opt backend builders to use typescript5
+  const useTypescript5 = considerBuildCommand;
   debug('Tracing input files...');
   const traceTime = Date.now();
   const { preparedFiles, shouldAddSourcemapSupport } = await compile(
@@ -515,7 +530,8 @@ export const build = async ({
     config,
     meta,
     nodeVersion,
-    isEdgeFunction
+    isEdgeFunction,
+    useTypescript5
   );
   debug(`Trace complete [${Date.now() - traceTime}ms]`);
 
@@ -607,6 +623,9 @@ export const build = async ({
       regions: normalizeRequestedRegions(
         staticConfig?.preferredRegion ?? staticConfig?.regions
       ),
+      shouldDisableAutomaticFetchInstrumentation:
+        process.env.VERCEL_TRACING_DISABLE_AUTOMATIC_FETCH_INSTRUMENTATION ===
+        '1',
     });
   }