+++ packages/oidc/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @vercel/oidc
 
+## 3.1.0
+
+### Minor Changes
+
+- Allow vercel/oidc to refresh the vercel CLI auth token when running locally ([#14543](https://github.com/vercel/vercel/pull/14543))
+
+### Patch Changes
+
+- improve error messages for package consumers ([#14449](https://github.com/vercel/vercel/pull/14449))
+
+## 3.0.5
+
+### Patch Changes
+
+- Fix OIDC token expiry check ([#14306](https://github.com/vercel/vercel/pull/14306))
+
+## 3.0.4
+
+### Patch Changes
+
+- Fix directory permissions so that files can be created under the OIDC data directory in linux ([#14214](https://github.com/vercel/vercel/pull/14214))
+
 ## 3.0.3
 
 ### Patch Changes